requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
pycryptodome=3.14.1
alembic~=1.7.5
SQLAlchemy~=1.4.29
paho-mqtt~=1.6.1